Add conditional checks to prevent NullPointerException

+++ backend/src/main/java/com/dsa/graphs/service/AdjacencyListServiceImpl.java
@@ -160,8 +160,10 @@ public class AdjacencyListServiceImpl implements AdjacencyListService {
 
     @Override
     public void deleteAdjacencyList() {
-        LOGGER.info("------ DELETING ADJACENCY LIST");
-        adjacencyList.delete();
-        LOGGER.info("------ SUCCESSFULLY DELETED ADJACENCY LIST");
+        if (adjacencyList != null) {
+            LOGGER.info("------ DELETING ADJACENCY LIST");
+            adjacencyList.delete();
+            LOGGER.info("------ SUCCESSFULLY DELETED ADJACENCY LIST");
+        }
     }
 }
